pyproject.toml
[project]
name = "bingobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.19.0",
    "discord-py>=2.3.0",
    "openai>=1.59.8",
    "openai-agents>=0.0.16",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.3", "pytest-asyncio>=0.25.0"]

[dependency-groups]
dev = ["pytest>=8.4.2", "pytest-asyncio>=1.2.0"]
